@@ -1,28 +1,30 @@
 name: flutter_cache_manager
 description: Generic cache manager for flutter. Saves web files on the storages of the device and saves the cache info using sqflite.
-version: 3.3.0
-homepage: https://github.com/Baseflow/flutter_cache_manager/tree/master/flutter_cache_manager
-
+version: 3.3.1
+homepage: https://github.com/Baseflow/flutter_cache_manager/tree/develop/flutter_cache_manager
+topics:
+  - cache
+  - cache-manager
 environment:
-  sdk: ">=2.12.0 <3.0.0"
+  sdk: '>=2.17.0 <4.0.0'
 
 dependencies:
   clock: ^1.1.0
   collection: ^1.15.0
-  file: ">=6.0.0 <7.0.0"
+  file: '>=6.1.4 <8.0.0'
   flutter:
     sdk: flutter
-  http: ^0.13.0
+  http: '>=0.13.0 <2.0.0'
   path: ^1.8.0
   path_provider: ^2.0.0
-  rxdart: ">=0.26.0 <0.28.0"
-  # sqflite: ^2.0.0+3
+  rxdart: '>=0.26.0 <0.28.0'
+  # sqflite: ^2.0.0+4
   sqflite_sqlcipher: ^2.1.1+1
-  uuid: ^3.0.0
+  uuid: '>=3.0.0 <5.0.0'
 
 dev_dependencies:
-  build_runner: ^1.11.5
+  build_runner: ^2.0.0
+  flutter_lints: ^2.0.1
   flutter_test:
     sdk: flutter
   mockito: ^5.0.0
-  flutter_lints: ^1.0.4
