@@ -1,35 +1,38 @@
 import 'package:baseflow_plugin_template/baseflow_plugin_template.dart';
 import 'package:example/plugin_example/download_page.dart';
 import 'package:example/plugin_example/floating_action_button.dart';
+import 'package:flutter/foundation.dart' show kDebugMode;
 import 'package:flutter/material.dart';
 import 'package:flutter_cache_manager/flutter_cache_manager.dart';
 
 void main() {
-  runApp(BaseflowPluginExample(
-    pluginName: 'Flutter Cache Manager',
-    githubURL: 'https://github.com/Baseflow/flutter_cache_manager',
-    pubDevURL: 'https://pub.dev/packages/flutter_cache_manager',
-    pages: [CacheManagerPage.createPage()],
-  ));
+  runApp(
+    BaseflowPluginExample(
+      pluginName: 'Flutter Cache Manager',
+      githubURL: 'https://github.com/Baseflow/flutter_cache_manager',
+      pubDevURL: 'https://pub.dev/packages/flutter_cache_manager',
+      pages: [CacheManagerPage.createPage()],
+    ),
+  );
   CacheManager.logLevel = CacheManagerLogLevel.verbose;
 }
 
-const url = 'https://blurha.sh/assets/images/img1.jpg';
+const url = 'https://picsum.photos/200/300';
 
 /// Example [Widget] showing the functionalities of flutter_cache_manager
 class CacheManagerPage extends StatefulWidget {
-  const CacheManagerPage({Key key}) : super(key: key);
+  const CacheManagerPage({super.key});
 
   static ExamplePage createPage() {
     return ExamplePage(Icons.save_alt, (context) => const CacheManagerPage());
   }
 
   @override
-  _CacheManagerPageState createState() => _CacheManagerPageState();
+  CacheManagerPageState createState() => CacheManagerPageState();
 }
 
-class _CacheManagerPageState extends State<CacheManagerPage> {
-  Stream<FileResponse> fileStream;
+class CacheManagerPageState extends State<CacheManagerPage> {
+  Stream<FileResponse>? fileStream;
 
   void _downloadFile() {
     setState(() {
@@ -41,16 +44,16 @@ class _CacheManagerPageState extends State<CacheManagerPage> {
   Widget build(BuildContext context) {
     if (fileStream == null) {
       return Scaffold(
-        appBar: null,
         body: const ListTile(
-            title: Text('Tap the floating action button to download.')),
+          title: Text('Tap the floating action button to download.'),
+        ),
         floatingActionButton: Fab(
           downloadFile: _downloadFile,
         ),
       );
     }
     return DownloadPage(
-      fileStream: fileStream,
+      fileStream: fileStream!,
       downloadFile: _downloadFile,
       clearCache: _clearCache,
       removeFile: _removeFile,
@@ -66,11 +69,13 @@ class _CacheManagerPageState extends State<CacheManagerPage> {
 
   void _removeFile() {
     DefaultCacheManager().removeFile(url).then((value) {
-      //ignore: avoid_print
-      print('File removed');
+      if (kDebugMode) {
+        print('File removed');
+      }
     }).onError((error, stackTrace) {
-      //ignore: avoid_print
-      print(error);
+      if (kDebugMode) {
+        print(error);
+      }
     });
     setState(() {
       fileStream = null;
