@@ -2,7 +2,7 @@ import 'dart:async';
 import 'dart:typed_data';
 
 import 'package:file/file.dart';
-import 'package:flutter/foundation.dart';
+import 'package:flutter/foundation.dart' show visibleForTesting;
 import 'package:flutter_cache_manager/flutter_cache_manager.dart';
 import 'package:flutter_cache_manager/src/cache_store.dart';
 import 'package:flutter_cache_manager/src/storage/cache_object.dart';
@@ -114,8 +114,13 @@ class CacheManager implements BaseCacheManager {
     return streamController.stream;
   }
 
-  Future<void> _pushFileToStream(StreamController streamController, String url,
-      String? key, Map<String, String>? headers, bool withProgress) async {
+  Future<void> _pushFileToStream(
+    StreamController<dynamic> streamController,
+    String url,
+    String? key,
+    Map<String, String>? headers,
+    bool withProgress,
+  ) async {
     key ??= url;
     FileInfo? cacheFile;
     try {
@@ -124,14 +129,14 @@ class CacheManager implements BaseCacheManager {
         streamController.add(cacheFile);
         withProgress = false;
       }
-    } catch (e) {
+    } on Object catch (e) {
       cacheLogger.log(
           'CacheManager: Failed to load cached file for $url with error:\n$e',
           CacheManagerLogLevel.debug);
     }
     if (cacheFile == null || cacheFile.validTill.isBefore(DateTime.now())) {
       try {
-        await for (var response
+        await for (final response
             in _webHelper.downloadFile(url, key: key, authHeaders: headers)) {
           if (response is DownloadProgress && withProgress) {
             streamController.add(response);
@@ -140,7 +145,7 @@ class CacheManager implements BaseCacheManager {
             streamController.add(response);
           }
         }
-      } catch (e) {
+      } on Object catch (e) {
         cacheLogger.log(
             'CacheManager: Failed to download file from $url with error:\n$e',
             CacheManagerLogLevel.debug);
@@ -149,7 +154,7 @@ class CacheManager implements BaseCacheManager {
         }
       }
     }
-    unawaited(streamController.close());
+    streamController.close();
   }
 
   ///Download the file and add to cache
@@ -159,7 +164,7 @@ class CacheManager implements BaseCacheManager {
       Map<String, String>? authHeaders,
       bool force = false}) async {
     key ??= url;
-    var fileResponse = await _webHelper
+    final fileResponse = await _webHelper
         .downloadFile(
           url,
           key: key,
@@ -213,7 +218,7 @@ class CacheManager implements BaseCacheManager {
 
     final file = await _config.fileSystem.createFile(cacheObject.relativePath);
     await file.writeAsBytes(fileBytes);
-    unawaited(_store.putFile(cacheObject));
+    _store.putFile(cacheObject);
     return file;
   }
 
@@ -246,16 +251,16 @@ class CacheManager implements BaseCacheManager {
       eTag: eTag,
     );
 
-    var file = await _config.fileSystem.createFile(cacheObject.relativePath);
+    final file = await _config.fileSystem.createFile(cacheObject.relativePath);
 
     // Always copy file
-    var sink = file.openWrite();
+    final sink = file.openWrite();
     await source
         // this map is need to map UInt8List to List<int>
         .map((event) => event)
         .pipe(sink);
 
-    unawaited(_store.putFile(cacheObject));
+    _store.putFile(cacheObject);
     return file;
   }
 
