@@ -1,4 +1,4 @@
-import '../../flutter_cache_manager.dart';
+import 'package:flutter_cache_manager/flutter_cache_manager.dart';
 
 /// The DefaultCacheManager that can be easily used directly. The code of
 /// this implementation can be used as inspiration for more complex cache
@@ -7,6 +7,7 @@ class DefaultCacheManager extends CacheManager with ImageCacheManager {
   static const key = 'libCachedImageData';
 
   static final DefaultCacheManager _instance = DefaultCacheManager._();
+
   factory DefaultCacheManager() {
     return _instance;
   }
