@@ -1,9 +1,8 @@
 import 'package:flutter_cache_manager/flutter_cache_manager.dart';
-import 'package:flutter_cache_manager/src/storage/file_system/file_system.dart';
-
-import '_config_unsupported.dart'
+import 'package:flutter_cache_manager/src/config/_config_unsupported.dart'
     if (dart.library.html) '_config_web.dart'
     if (dart.library.io) '_config_io.dart' as impl;
+import 'package:flutter_cache_manager/src/storage/file_system/file_system.dart';
 
 abstract class Config {
   /// Config file for the CacheManager.
@@ -32,9 +31,14 @@ abstract class Config {
   }) = impl.Config;
 
   String get cacheKey;
+
   Duration get stalePeriod;
+
   int get maxNrOfCacheObjects;
+
   CacheInfoRepository get repo;
+
   FileSystem get fileSystem;
+
   FileService get fileService;
 }
