@@ -1,11 +1,11 @@
 import 'dart:io';
 
+import 'package:flutter_cache_manager/src/storage/cache_info_repositories/cache_info_repository.dart';
 import 'package:flutter_cache_manager/src/storage/cache_info_repositories/helper_methods.dart';
+import 'package:flutter_cache_manager/src/storage/cache_object.dart';
 import 'package:path/path.dart';
 import 'package:path_provider/path_provider.dart';
 import 'package:sqflite_sqlcipher/sqflite.dart';
-import '../cache_object.dart';
-import 'cache_info_repository.dart';
 
 const _tableCacheObject = 'cacheObject';
 
@@ -25,7 +25,7 @@ class CacheObjectProvider extends CacheInfoRepository
     if (!shouldOpenOnNewConnection()) {
       return openCompleter!.future;
     }
-    var path = await _getPath();
+    final path = await _getPath();
     await File(path).parent.create(recursive: true);
     db = await openDatabase(path, version: 3,
         onCreate: (Database db, int version) async {
@@ -98,7 +98,7 @@ class CacheObjectProvider extends CacheInfoRepository
   @override
   Future<CacheObject> insert(CacheObject cacheObject,
       {bool setTouchedToNow = true}) async {
-    var id = await db!.insert(
+    final id = await db!.insert(
       _tableCacheObject,
       cacheObject.toMap(setTouchedToNow: setTouchedToNow),
     );
@@ -107,7 +107,7 @@ class CacheObjectProvider extends CacheInfoRepository
 
   @override
   Future<CacheObject?> get(String key) async {
-    List<Map> maps = await db!.query(_tableCacheObject,
+    final List<Map<dynamic, dynamic>> maps = await db!.query(_tableCacheObject,
         columns: null, where: '${CacheObject.columnKey} = ?', whereArgs: [key]);
     if (maps.isNotEmpty) {
       return CacheObject.fromMap(maps.first.cast<String, dynamic>());
@@ -124,7 +124,7 @@ class CacheObjectProvider extends CacheInfoRepository
   @override
   Future<int> deleteAll(Iterable<int> ids) {
     return db!.delete(_tableCacheObject,
-        where: '${CacheObject.columnId} IN (' + ids.join(',') + ')');
+        where: '${CacheObject.columnId} IN (${ids.join(',')})');
   }
 
   @override
@@ -178,7 +178,7 @@ class CacheObjectProvider extends CacheInfoRepository
   }
 
   @override
-  Future deleteDataFile() async {
+  Future<void> deleteDataFile() async {
     await _getPath();
   }
 
@@ -193,7 +193,7 @@ class CacheObjectProvider extends CacheInfoRepository
     if (_path != null) {
       directory = File(_path!).parent;
     } else {
-      directory = (await getApplicationSupportDirectory());
+      directory = await getApplicationSupportDirectory();
     }
     await directory.create(recursive: true);
     if (_path == null || !_path!.endsWith('.db')) {
@@ -204,8 +204,8 @@ class CacheObjectProvider extends CacheInfoRepository
   }
 
   // Migration for pre-V2 path on iOS and macOS
-  Future _migrateOldDbPath(String newDbPath) async {
-    final oldDbPath = join((await getDatabasesPath()), '$databaseName.db');
+  Future<void> _migrateOldDbPath(String newDbPath) async {
+    final oldDbPath = join(await getDatabasesPath(), '$databaseName.db');
     if (oldDbPath != newDbPath && await File(oldDbPath).exists()) {
       try {
         await File(oldDbPath).rename(newDbPath);
