@@ -1,6 +1,5 @@
+import 'package:flutter_cache_manager/src/storage/cache_info_repositories/cache_info_repository.dart';
 import 'package:flutter_cache_manager/src/storage/cache_object.dart';
-
-import 'cache_info_repository.dart';
 
 class NonStoringObjectProvider implements CacheInfoRepository {
   @override
@@ -20,7 +19,7 @@ class NonStoringObjectProvider implements CacheInfoRepository {
 
   @override
   Future<CacheObject?> get(String url) {
-    return Future.value(null);
+    return Future.value();
   }
 
   @override
